integreted asset api

+++ src/main/java/com/chirag/service/OrderServiceImpl.java
@@ -2,10 +2,7 @@ package com.chirag.service;
 
 import com.chirag.domain.OrderStatus;
 import com.chirag.domain.OrderType;
-import com.chirag.modal.Coin;
-import com.chirag.modal.Order;
-import com.chirag.modal.OrderItem;
-import com.chirag.modal.User;
+import com.chirag.modal.*;
 import com.chirag.repository.OrderItemRepository;
 import com.chirag.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,6 +24,9 @@ public class OrderServiceImpl implements OrderService{
 
     @Autowired
     private OrderItemRepository orderItemRepository;
+
+    @Autowired
+    private AssetService assetService;
 
     @Override
     public Order createOrder(User user, OrderItem orderItem, OrderType orderType) {
@@ -86,6 +86,14 @@ public class OrderServiceImpl implements OrderService{
 
         //create asset
 
+        Asset oldAsset = assetService.findAssetByUserIdAndCoinId(order.getUser().getId(), order.getOrderItem().getCoin().getId());
+
+        if(oldAsset ==null){
+            assetService.createAsset(user, orderItem.getCoin(), orderItem.getQuantity());
+        }else{
+            assetService.updateAsset(oldAsset.getId(), quantity);
+        }
+
         return savedOrder;
     }
 
@@ -96,28 +104,39 @@ public class OrderServiceImpl implements OrderService{
         }
         double sellPrice = coin.getCurrentPrice();
 
-        double buyPrice = assetToSell.getPrice();
+        Asset assetToSell = assetService.findAssetByUserIdAndCoinId(user.getId(), coin.getId());
+        double buyPrice = assetToSell.getBuyPrice();
 
-        OrderItem orderItem = createOrderItem(coin, quantity,buyPrice,sellPrice);
+        if(assetToSell != null){
 
-        Order order = createOrder(user, orderItem, OrderType.SELL);
-        orderItem.setOrder(order);
 
-        if(assetToSell.getQuantity()>=quantity){
-            order.setStatus(OrderStatus.SUCCESS);
-            order.setOrderType(OrderType.SELL);
-            Order savedOrder = orderRepository.save(order);
+            OrderItem orderItem = createOrderItem(coin, quantity,buyPrice,sellPrice);
 
-            walletService.payOrderPayment(order,user);
+            Order order = createOrder(user, orderItem, OrderType.SELL);
+            orderItem.setOrder(order);
 
-            Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
-            if(updatedAsset.getQuantity()*coin.getCurrentPrice() <=1){
-                assetService.deleteAsset(updatedAsset.getId());
+            if(assetToSell.getQuantity()>=quantity){
+                order.setStatus(OrderStatus.SUCCESS);
+                order.setOrderType(OrderType.SELL);
+                Order savedOrder = orderRepository.save(order);
+
+                walletService.payOrderPayment(order,user);
+
+                Asset updatedAsset = assetService.updateAsset(
+                        assetToSell.getId(), -quantity
+                );
+
+
+                if(updatedAsset.getQuantity()*coin.getCurrentPrice() <=1){
+                    assetService.deleteAsset(updatedAsset.getId());
+                }
+                return savedOrder;
             }
-            return savedOrder;
+
+            throw new Exception("Insufficient Quantity to sell");
         }
 
-        throw new Exception("Insufficient Quantity to sell");
+        throw new Exception("Asset not found");
 
     }
